SOAR/tag_triage_agent/requirements.txt
requests==2.31.0
python-dotenv==0.21.1
google-auth==2.23.3